+++ OpenHD/ohd_interface/CMakeLists.txt
@@ -14,19 +14,19 @@ set(sources
     "inc/ohd_interface.h"
     "inc/ohd_interface_settings.hpp"
     "inc/wifi_card.hpp"
-    "inc/USBTetherListener.h"
+    "inc/usb_tether_listener.h"
     "inc/validate_settings_helper.h"
     "inc/WBStreams.h"
     "inc/WBStreamsSettings.hpp"
-    "inc/WifiHotspot.h"
+    "inc/wifi_hotspot.h"
     #"inc/wifi_command_helper2.h"
 
     "src/wifi_card_discovery.cpp"
     "src/ohd_interface.cpp"
-    "src/USBTetherListener.cpp"
+    "src/usb_tether_listener.cpp"
     "src/WBStreams.cpp"
-    "src/WifiCardCommandHelper.hpp"
-    "src/WifiHotspot.cpp"
+    "src/wifi_command_helper.hpp"
+    "src/wifi_hotspot.cpp"
     #"src/wifi_command_helper2.cpp"
 )
