+++ OpenHD/ohd_interface/inc/wifi_hotspot.h
