+++ OpenHD/ohd_interface/src/wifi_command_helper.hpp
