+++ OpenHD/ohd_interface/src/wifi_hotspot.cpp
@@ -2,11 +2,10 @@
 // Created by consti10 on 17.05.22.
 //
 
-#include "WifiHotspot.h"
-
-#include "openhd-spdlog.hpp"
 #include <utility>
 
+#include "openhd-spdlog.hpp"
+#include "wifi_hotspot.h"
 
 WifiHotspot::WifiHotspot(WiFiCard wifiCard):
 wifiCard(std::move(wifiCard)) {
