+++ OpenHD/ohd_interface/test/test_usb_hotspot.cpp
@@ -2,8 +2,8 @@
 // Created by consti10 on 20.05.22.
 //
 
-#include "USBTetherListener.h"
 #include "openhd-util.hpp"
+#include "usb_tether_listener.h"
 
 int main(int argc, char *argv[]) {
